chore(ci): stand up initial ci pipeline
this commit introduces a basic ci pipeline for the project. `main.yml`
acts as the entrypoint for the pipeline, which can be called in one of
three ways:
1. as a part of a merge group, which is already setup for the repo
2. on push to `main`, which we've already protected properly
3. on any pull_request, which ought to be fine since there are no
secrets in this repo at the moment

the `main` entrypoint only calls `build` at this time, which downloads
the project's dependencies and builds the project.