chore(ci): add initial testing ci pipeline
add an initial testing pipeline to ci. at the time of this writing, the
project doesn't have any formal testing infra set up. as a result, we
run a stub command defined in `package.json` that just echoes some
nonesense to the terminal in the interim.

the pipeline is updated such that the build workflow uploads the build
archive for a new test-unit workflow to pull down and run (however,
today we're not actually using it, this just puts the plumbing in
place).

to accomplish this, new actions for uploading/downloading the archive
that match stencil core are added to the project