chore(ci): stand up initial ci pipeline (#6)
* chore(ci): stand up initial ci pipeline

this commit introduces a basic ci pipeline for the project. `main.yml`
acts as the entrypoint for the pipeline, which can be called in one of
three ways:
1. as a part of a merge group, which is already setup for the repo
2. on push to `main`, which we've already protected properly
3. on any pull_request, which ought to be fine since there are no
secrets in this repo at the moment

the `main` entrypoint only calls `build` at this time, which downloads
the project's dependencies and builds the project.

* fix(build): error related to "NodeNext" module resolution

This commit reverts a change made to `module` and `moduleResolution` in the `tsconfig.json` file to set them to "NodeNext". This conflicts with the module resolution strategy in Stencil core dependency.

---------

Co-authored-by: Tanner Reits <tanner@ionic.io>