chore(repo): add automated dev releases
add the ability to publish a dev build of the project to the npm
registry. we do so by adding a new workflow that can be manually kicked
off, `release-dev`. as the name implies, it creates a dev build (running
the most basic/essential parts of the build pipeline) and invokes a
publish to npm