chore(repo): stub production releases
this commit creates the initial pass at automated production releases
for the repo. it is capable of:
1. building stencil playwright
2. versioning the library based on user input
3. logging the generated change to the console
4. generating and pushing a commit to _a_ branch

the intent of this commit is to validate some of the mechanics of user
input before we start attempting to publish. the plan is for that to
come in a later pr.

unknown: I _think_ our branch protection rules will hinder
us from pushing to `main` once we try to do this "for real". we had
some difficulty here with stencil-core, and may have it again here.
we'll merge this to verify the input works as expected. we can try to
derisk this by setting up similar branch protection rules on a dummy
branch and proceed from there