chore(repo): pivot releases to creating a pr
as mentioned in https://github.com/ionic-team/stencil-playwright/pull/15,
it turns out our branch protections will prevent us from pushing to
`main` in a github action.

in the interest of time and to provide some release consistency to the
rest of the team, this commmit pivots the existing `release-prod`
workflow to a workflow that will only create a pull request that bumps
the version of the project.

in a separate commit/pr, we shall introduce a new workflow to do the
actual publish