chore(repo): create production release workflow
note: this is an initial pr that does not enable production publishes.
read on to understand what this commit does and does not do.

add the basis of a produciton release github action workflow. this
workflow allows a use to provide a `tag` for npm to publish a new
version of package under.

this workflow builds the package, pulls the current version string out
of `package.json` (it assumes that the `create-production-pr` workflow
has already run), and prints it to the console.

this workflow does not yet invoke the publish-npm action. the reason
for this is safety/preventing unwanted publishes while this workflow is
created. since `releae-production` has not landed in the codebase,
attempting to run `publish-npm` would pull the action from `main`. this
is undesirable, as there are a handful of changes/modifications we
wish to make to the workflow to test `release-produciton` with regard
to tag publishes to github, not bumping the version in npm, etc.

instead, we opt to 'stub' this workflow for now to land it, then we can
safely iterate on it using a non-`push` target